requirements.txt
django ==4.2
djangorestframework ==3.14.0
django-cors-headers ==4.7.0
pandas ==2.3.2
openpyxl==3.1.2
